Move encode/abi to abi/encode
