+++ packages/codec/lib/abi/index.ts
@@ -1,7 +1,7 @@
 import * as Allocate from "./allocate";
 export { Allocate };
 
-import * as Encode from "@truffle/codec/encode/abi";
+import * as Encode from "./encode";
 export { Encode };
 
 export * from "./types";
